fix(ci): python-coverage-shards — split oversized test_executor_coverage_edges_part_003
Near-miss plan-recovery work grew part_003 to 1585 lines, tripping the
test_first_party_code_files_stay_under_line_limit maintainability guard
(1500-line cap). Move the four self-contained execution-validation tests
into a new part_015 file and prune the now-unused imports from part_003.
Behavior and coverage are preserved; both files now sit under the limit.
